requirements.txt
pytest
pytest-playwright
playwright
allure-pytest
faker
pytest-rerunfailures
